LibDebug: Move Dwarf::AttributeValue to a separate file
